fully symfony; also claude
