+++ CLAUDE.md
@@ -1,0 +1,166 @@
+# CLAUDE.md
+
+This file provides guidance to Claude Code (claude.ai/code) when working with code in this repository.
+
+## Project Overview
+
+OpenEMR is a Free and Open Source electronic health records and medical practice management application. This is a PHP-based web application with a complex architecture supporting healthcare workflows including patient management, billing, scheduling, and clinical documentation.
+
+## Architecture
+
+- **Backend**: PHP 8.2+ with MySQL/MariaDB database
+- **Frontend**: Mix of legacy PHP templates, jQuery, and modern JavaScript components
+- **Framework**: Hybrid architecture using both Symfony (Console, DI, Config, HttpFoundation) and Laminas components, with the project moving toward Symfony for new features
+- **API**: REST API and FHIR R4 compliant endpoints in `/src/RestControllers/` and `/src/FHIR/`
+- **Module System**: Extensible module architecture in `/modules/` directory
+- **Theme System**: SCSS-based themes in `/interface/themes/`
+
+### Current vs. Symfony Standard Structure
+
+**Current OpenEMR Structure:**
+```
+/src/                          # Modern PSR-4 classes
+/interface/                    # Legacy web interface
+/library/                      # Legacy utilities
+/public/                       # Public assets
+/templates/                    # Twig templates (limited use)
+/config/                       # Configuration files
+```
+
+**Symfony Standard Structure (target for gradual migration):**
+```
+/src/
+  /Controller/                 # HTTP controllers
+  /Entity/                     # Doctrine entities
+  /Repository/                 # Database repositories
+  /Service/                    # Business logic services
+  /Command/                    # Console commands
+/templates/                    # Twig templates
+/public/                       # Web root with index.php
+/config/                       # Configuration (services.yaml, routes.yaml)
+/var/                          # Cache, logs, sessions
+/migrations/                   # Database migrations
+```
+
+### Key Directories
+
+- `/src/` - Modern PSR-4 autoloaded PHP classes (Services, Controllers, etc.)
+- `/library/` - Legacy PHP libraries and utilities
+- `/interface/` - Web interface files (forms, reports, admin pages)
+- `/public/` - Public web assets
+- `/sql/` - Database schema and upgrade scripts
+- `/tests/` - PHPUnit test suite
+- `/docker/` - Docker configurations for development
+
+## Development Commands
+
+### Setup
+```bash
+composer install --no-dev
+npm install
+npm run build
+composer dump-autoload -o
+```
+
+### Console Commands (Symfony)
+```bash
+bin/console                                    # List all available commands
+bin/console openemr:acl-modify                # Modify ACL permissions
+bin/console openemr:register                  # Register Zend modules
+bin/console openemr:zfc-module                # Module management
+bin/console openemr:ccda-import               # Import CCDA documents
+bin/console openemr:create-api-documentation  # Generate API docs
+```
+
+### Frontend Development
+```bash
+npm run dev              # Build and watch for changes
+npm run gulp-build       # Build assets
+npm run gulp-watch       # Watch for changes
+```
+
+### Code Quality
+```bash
+# PHP
+composer phpstan         # Static analysis
+composer rector          # Code modernization
+vendor/bin/phpunit       # Run PHP tests
+
+# JavaScript
+npm run lint:js          # ESLint
+npm run lint:js-fix      # Fix ESLint issues
+npm run test:js          # Jest tests
+npm run test:js-coverage # Jest with coverage
+
+# CSS
+npm run stylelint        # Style linting
+npm run stylelint-fix    # Fix style issues
+```
+
+## Testing
+
+- **PHP Tests**: PHPUnit configuration in `phpunit.xml`
+- **JavaScript Tests**: Jest configuration in `jest.config.js`
+- **Test Location**: `/tests/` directory
+- **Database**: Tests use separate test database schemas
+
+## Code Standards
+
+### AI-Generated Code Requirements
+When contributing AI-generated code:
+- Add comment blocks marking AI-generated sections
+- Include line-end comments for single AI-generated lines
+- Note AI assistance in documentation/comments
+- Be specific about which parts were AI-generated
+- Indicate extent of modifications to AI code
+
+### PHP Standards
+- PSR-4 autoloading for new classes in `/src/`
+- Use Symfony Console for new CLI commands (extend `Symfony\Component\Console\Command\Command`)
+- Symfony Dependency Injection container preferred for new services
+- Database access through Services pattern in `/src/Services/`
+- Input validation using Particle\Validator or similar
+
+### Migration Strategy (Gradual Symfony Adoption)
+- **Controllers**: Move new controllers to `/src/Controller/` following Symfony patterns
+- **Services**: Continue building services in `/src/Services/` with DI
+- **Commands**: All new CLI commands use Symfony Console in `/src/Command/`
+- **Templates**: Gradually migrate to Twig templates for new features
+- **Configuration**: Move toward Symfony-style configuration files
+- **Legacy Support**: Maintain `/interface/` and `/library/` for existing functionality
+
+### Security Considerations
+- This is healthcare software - maintain HIPAA compliance
+- Sanitize all user inputs
+- Use prepared statements for database queries
+- Implement proper access controls and audit logging
+
+## Database
+
+- **Main Config**: `/sites/default/sqlconf.php`
+- **Schema**: `/sql/database.sql`
+- **Upgrades**: Version-specific upgrade scripts in `/sql/`
+- **Legacy Tables**: Many legacy table structures exist
+- **Modern Approach**: Use Laminas\Db for new database code
+
+## Module Development
+
+- **Location**: `/modules/` directory
+- **Structure**: Each module has its own directory with standard structure
+- **Registration**: Modules register through the module system
+- **API**: Modules can extend REST endpoints and add UI components
+
+## Contributing Guidelines
+
+### Code Areas to Avoid
+- **`/contrib/` directory**: Generally do not maintain or update files in this directory
+- **Exception**: `/contrib/util/installScripts/InstallerAuto.php` - This is actively maintained as the main installer
+
+The `contrib/` directory contains community contributions, utilities, and legacy tools that are not part of the core codebase maintenance.
+
+## Common Pitfalls
+
+- Mixed legacy and modern code patterns - follow existing patterns in the area you're working
+- Complex permission system - check existing ACL patterns
+- Multiple configuration systems - globals.php, site-specific configs
+- Internationalization required - use translation functions for user-facing text
