+++ ci/ciLibrary.source
@@ -107,7 +107,7 @@ phpcov() {
 install_configure() {
     actions_chmod 0666 sites/default/sqlconf.php
     actions_chmod -R 0777 sites/default/documents
-    _exec php contrib/util/installScripts/InstallerAuto.php --rootpass=root --server=mysql --loginhost=%
+    _exec bin/console openemr:install --rootpass=root --server=mysql --loginhost=%
     _exec mysql -u openemr --password="openemr" --ssl=false -h mysql -e 'INSERT INTO product_registration (opt_out) VALUES (1)' openemr
     _exec mysql -u openemr --password="openemr" --ssl=false -h mysql -e 'UPDATE globals SET gl_value = 1 WHERE gl_name = "rest_api"' openemr
     _exec mysql -u openemr --password="openemr" --ssl=false -h mysql -e 'UPDATE globals SET gl_value = 1 WHERE gl_name = "rest_fhir_api"' openemr
